Moved mocks into a common.test_mocks package, re #24475
