+++ qt/python/mantidqt/widgets/common/test/test_workspacedisplay_ads_observer.py
@@ -1,3 +1,12 @@
+# Mantid Repository : https://github.com/mantidproject/mantid
+#
+# Copyright &copy; 2019 ISIS Rutherford Appleton Laboratory UKRI,
+#     NScD Oak Ridge National Laboratory, European Spallation Source
+#     & Institut Laue - Langevin
+# SPDX - License - Identifier: GPL - 3.0 +
+#  This file is part of the mantid workbench.
+
+
 from __future__ import (absolute_import, division, print_function)
 
 import unittest
