+++ qt/python/mantidqt/widgets/matrixworkspacedisplay/test/test_matrixworkspacedisplay_model.py
@@ -16,7 +16,7 @@ from mock import Mock
 from mantid.simpleapi import CreateSampleWorkspace
 from mantidqt.widgets.matrixworkspacedisplay.model import MatrixWorkspaceDisplayModel
 from mantidqt.widgets.matrixworkspacedisplay.table_view_model import MatrixWorkspaceTableViewModelType
-from mantidqt.widgets.matrixworkspacedisplay.test_helpers.matrixworkspacedisplay_common import \
+from mantidqt.widgets.common.test_mocks.mock_mantid import \
     MockWorkspace
