+++ qt/python/mantidqt/widgets/tableworkspacedisplay/test/test_tableworkspacedisplay_model.py
@@ -15,8 +15,7 @@ import unittest
 from mock import Mock
 
 from mantid.kernel import V3D
-from mantidqt.widgets.matrixworkspacedisplay.test_helpers.matrixworkspacedisplay_common import \
-    MockWorkspace
+from mantidqt.widgets.common.test_mocks.mock_mantid import MockWorkspace
 from mantidqt.widgets.tableworkspacedisplay.model import TableWorkspaceDisplayModel
